[#5] Extract common DSL elements
