+++ accurest-core/src/main/groovy/io/coderate/accurest/dsl/GroovyDsl.groovy
@@ -1,11 +1,13 @@
 package io.coderate.accurest.dsl
 
 import groovy.transform.TypeChecked
+import io.coderate.accurest.dsl.internal.Response
 
 @TypeChecked
 class GroovyDsl {
 
     Request request
+    Response response
 
     static GroovyDsl make(Closure closure) {
         GroovyDsl dsl = new GroovyDsl()
@@ -18,6 +20,13 @@ class GroovyDsl {
         Request request = new Request()
         this.request = request
         closure.delegate = request
+        closure()
+    }
+
+    void response(@DelegatesTo(Response) Closure closure) {
+        Response response = new Response()
+        this.response = response
+        closure.delegate = response
         closure()
     }
 }
