+++ accurest-core/src/main/groovy/io/coderate/accurest/dsl/internal/DelegateHelper.groovy
@@ -1,0 +1,10 @@
+package io.coderate.accurest.dsl.internal
+
+//TODO: There is problem with a trait usage
+class DelegateHelper {
+
+    public static <T> void delegateToClosure(@DelegatesTo(T) Closure closure, T delegate) {
+        closure.delegate = delegate
+        closure()
+    }
+}
