+++ accurest-core/src/main/groovy/io/coderate/accurest/dsl/internal/JSONCompareMode.groovy
@@ -1,0 +1,5 @@
+package io.coderate.accurest.dsl.internal
+
+enum JSONCompareMode {
+    STRICT, LENIENT, NON_EXTENSIBLE, STRICT_ORDER;
+}
